requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
httpx==0.27.2
pydantic==2.9.2
python-dotenv==1.0.1
motor
passlib[bcrypt]
bcrypt==4.0.1
